Move OAuth helpers to a different file
